refactor(parser): split primary/container.rs into cohesive submodules (§7-8)
Split the 1887-line parser/primary/container.rs (over the 500-line limit) into
a facade plus six themed submodules under src/parser/primary/container/. All
files are under 500 lines. container.rs is addressed as container::<name> (14
call sites in primary/mod.rs plus crate-wide path uses); the facade re-exports
each submodule at its original visibility, so no external caller changes.
primary/mod.rs is untouched.

- container.rs (34)       - facade: mod decls, re-exports
- allomorph.rs (290)      - allomorphic value construction from angle-word strings
- angle_words.rs (371)    - angle/french/double-angle quote-word list parsing
- array.rs (220)          - array composer [..], %(..), error helpers
- meta_ops.rs (492)       - zip/meta-op normalization + paren sequence/modifier helpers
- paren.rs (378)          - core parenthesized expression / list parsing
- sigil_context.rs (183)  - itemized/context forms: $(), @(), %(), $[], ${}

Pure relocation. Moved fns are pub(crate) in submodules and re-exported at their
original visibility; cross-submodule helpers bumped to pub(crate). Body refs to
primary-level siblings rewritten to absolute crate::parser paths. The fn multiset
is byte-identical. No behavior change.

External container:: callers preserved without edits:
- src/parser/primary/mod.rs (container:: prefix calls, untouched)
- src/parser/primary/misc/colonpair.rs (crate:: path, untouched)
- src/parser/primary/misc/hash.rs (crate:: path, untouched)
- src/parser/primary/string/quoted.rs (crate:: path, untouched)
- src/parser/primary/string/q_string.rs (crate:: path, untouched)
- src/parser/primary/var.rs (use super::container::paren_expr, untouched)
- src/runtime/builtins_string.rs (crate:: path, untouched)
